Put domain for voice memo row in a different file.
